requirements.txt
# Core Framework
Flask==2.3.3
Werkzeug==3.0.1
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0

# API & Validation
pydantic==2.5.0
flask-cors==4.0.0
flask-restx==1.3.0
marshmallow==3.20.1

# Authentication & Security
PyJWT==2.8.0
bcrypt==4.1.2
flask-jwt-extended==4.6.0

# Database & Caching
alembic==1.13.1
redis==5.0.1

# Background Tasks
celery==5.3.4

# Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Production
gunicorn==21.2.0
prometheus-client==0.19.0

# Development
black==23.12.0
flake8==6.1.0
pre-commit==3.6.0
mypy==1.7.1

# Utilities
click==8.1.7
python-dateutil==2.8.2
email-validator==2.1.0